requirements.txt
textual
groq
python-dotenv